% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\name{Search_online}
\alias{Search_online}
\title{Buscar series}
\usage{
Search_online(PATTERN = "*")
}
\arguments{
\item{PATTERN}{Pattern de búsqueda en la descripción de la serie}
}
\value{
Tibble con las series disponibles que con descripción coincidente
}
\description{
Buscar en el archivo de meta-datos online del Portal de Hacienda.
Utilizar \code{Search_online("*")} para descargar todos los metadatos y hacer búsquedas posteriores
sin descargar toda la base nuevamente.
}
\examples{
\donttest{
Listado <- Search_online("Tipo de Cambio")
Todaslasseries <- Search_online("*")
}
}
