% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncionesAcceso.R
\name{vForecast}
\alias{vForecast}
\title{Extender series con proyecciones de auto.arima (paquete Forecast)}
\usage{
vForecast(SERIE, N = 6, ...)
}
\arguments{
\item{SERIE}{XTS a extender}

\item{N}{Cantidad de períodos a extender (detecta automáticamente la frecuencia de la serie)}

\item{...}{Otros parámetros para \code{auto.arima}}
}
\value{
XTS con la series expandidas, acepta xts con muchas series
}
\description{
Recomendado sólo para estimaciones rápidas. A diferencia de \code{Forecast}, no devuelve intervalos de confianza, pero acepta
como input un XTS con múltiples series de tiempo.
}
\examples{
# Forecast de 12 meses del tipo de cambio
TCN <- vForecast(Get("120.1_PCE_1993_0_24,120.1_ED1_1993_0_26"),12)
}
